Change worker_processes value to auto
The nginx documentation says (http://nginx.org/en/docs/ngx_core_module.html#worker_processes): The optimal value depends on many factors including (but not limited to) the number of CPU cores, the number of hard disk drives that store data, and load pattern. When one is in doubt, setting it to the number of available CPU cores would be a good start (the value “auto” will try to autodetect it).

So changing the "worker_processes" value to auto is more general, then the current option.