Allow getting a cert for the main domain with DuckDNS
This commit is a follow up for #232, which changed DuckDNS to only support wildcards (*.mysubdomain.duckdns.org).

This commit tries to mimic the old behavior before #232 with additional support for wildcards. Ie. if SUBDOMAINS is set to 'wildcard' as instructed, a wildcard cert is requested as currently. Otherwise certificate will be for the main domain as it used to be before #232.